First step to to making finding anagrams faster
Added sorted words column to database.  Anagrams are found by sorting
the characters in the word into alphabetical order. Presorting the
words in the dictionary will make it faster to find anagrams.

Additionally, the sorted words column is an index. This makes finding
the anagrams a log(n) operation instead of an (n) operation.